[0.3.5]  fs  Admin delete API infinite loop bug fix test; decrease reply-count after deletion

+++ gesipan-fs/app/api/admin/del/[g]/route.ts
@@ -3,50 +3,78 @@ import type{NextRequest}from"next/server"
 import{mainDB,cmtDB}from"@/_lib/_insu"
 import{DEL}from"@/_lib/conf"
 
-export async function POST(req:NextRequest,{params}:{params:{g:string}}){
+export async function GET(req:NextRequest,{params}:{params:{g:string}}){
   const
     cmt = req.nextUrl.searchParams.get('cmt'),
     g = params.g
 
 // DELETE A POST
-  if(!cmt){
-    try{
-      const
-        resMain = await(await mainDB)
-          .deleteOne({g}),
-        resCmt = await(await cmtDB)
-          .deleteMany({g}),
-        pro='ADMIN_DEL_GESIMUL'
-      if(resMain.deletedCount===1){
-        console.log(`[${_t.t2+pro} 201] Deleted docu [g::${g}] and its ${resCmt} replies.`)
-        return NJ({del:1, delCmtCount:resCmt.deletedCount} as DEL)
-      }
-      _t.t500(pro)
-      return _t.NJ500
-    }catch(_){
-      _t.t500('ADMIN_DEL_GESIMUL_UNEXPECTED')
-      return _t.NJ500
+if(cmt===null){
+  try{
+    const
+      resMain = await(await mainDB)
+        .deleteOne({g}),
+      resCmt = await(await cmtDB)
+        .deleteMany({g}),
+      pro='ADMIN_DEL_GESIMUL'
+    if(resMain.deletedCount===1){
+      console.log(`[${_t.t2+pro} 201] Deleted docu [g::${g}] and its ${resCmt} replies.`)
+      return NJ({del:1, delCmtCount:resCmt.deletedCount} as DEL)
     }
+    _t.t500(pro)
+    console.log(await(await mainDB).find({g}).toArray())
+    console.error('Deleting ID: '+g)
+    console.error('Deleted doc: '+resMain.deletedCount)
+    console.log('Deleted cmd: '+resCmt.deletedCount)
+    return _t.NJ500
+  }catch(_){
+    _t.t500('ADMIN_DEL_GESIMUL_UNEXPECTED')
+    console.error('Caught:\n'+_)
+    return _t.NJ500
   }
+}
 
 // DELETE A CMT
+if(cmt==='1'){
   const
+
+    /** Flag obj: count the deletion of descendant replies. */
     delCount = {count:0,grand:false},
+
     /** Delete the reply and all its descendants. */
     delCmtDesc = async(no:string)=>{
-      await(await cmtDB)
-        .deleteOne({no})
-      const children = await(await cmtDB)
-        .find({tocmt:no}).toArray()
+    // minus
+      const self = await(await cmtDB).findOne({no})
+      if(!self) return _t.t422('ADMIN_DELETE_API_02')
+    // delete `no`
+      const
+        del = await(await cmtDB)
+          .deleteOne({no}),
+        children = await(await cmtDB)
+          .find({tocmt:no}).toArray()
+      del.deletedCount &&
+        await(await cmtDB).updateOne(
+          {no: self.tocmt},
+          {$inc: {ctc_count: -1}}
+        )
+    // delete descendants
       delCount.grand ? delCount.count++ : delCount.grand=true
-      for(const child of children) delCmtDesc(child.tocmt)
+      if(children.length===0) return delCount.count
+      for(const child of children) delCmtDesc(child.no)
     }
+
+///////
   try{
-    await delCmtDesc(g)
+    console.log(await delCmtDesc(g))
     console.log(`[${_t.t2+'ADMIN_DEL_REPLY'} 201] Deleted cmt [no::${g}] and its ${delCount.count} descendant replies.`)
     return NJ({del:1, delCmtCount: delCount.count} as DEL)
   }catch(_){
     _t.t500('ADMIN_DEL_GESIMUL_UNEXPECTED')
+    console.error('Caught:\n'+_)
     return _t.NJ500
   }
+}
+
+// 422
+  else{return _t.t422('ADMIN_DELETE_API_01')}
 }
